Chapter_5/requirements.txt
streamlit==1.38.0
boto3==1.34.162
langchain
langchain-community
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
